refactor: ì코드 리뷰 수정내용 ì»반영 완료

+++ src/main/java/lotto/domain/Prize.java
@@ -49,7 +49,7 @@ public enum Prize {
         return this;
     }
 
-    public int sumMoney(int totalMoney) {
+    public long sumMoney(long totalMoney) {
         return money + totalMoney;
     }
